Applied solidSession -> sessionInfo refactor
